Update patch set 5
Patch Set 5:

So this patch appears to work at this point (except for loopfilter TODO). I'm (ATM) getting losses overall on both derf and std-hd, but some files are already going up in both sets (e.g. ped and crew by about 5% in the low-med bitrate segment).

"Why is the rest going down?" Well, the reason for that is probably the same as the unit test failing right now: to fit it all in 16 bit intermediates, and the scale of our transforms being 8bit input + 1bit sign + 2bit precision (taken away after quantization) + 1bit per 2^n transform size (e.g. 1 for 2x2, 2 for 4x4, 3 for 8x8, 4 for 16x16, 5 for 32x32 and 6 for 64x64) + 1 bit extra that I don't know what it's for = 17bit for 32x32 and 18bit for 64x64.

Since we try to store everything in int16_t, I had to chop off 1 bit for 32x32 which I can restore after quantization. The same was done for 64x64, but now it's 2bits. After removing these 2 bits, overflows in quant and dct are gone, but precision is hurt a lot, and I get about 2-3 off-by-1s per 64x64 block (the unit test requires <0.1). As a quick-hack, I could probably protect against this by using a RD penalty against the tx64x64 to prevent it from being used when it's not useful, or doing the RD distortion in pixel-domain instead of coefficient-domain.

Or we could just not use the 64x64tx this time around. What do you guys think?

Patch-set: 5
Reviewer: Gerrit User 1000885 <1000885@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0
Label: Verified=0