Adds an error-resilient mode.
WIP: needs work on resolving a mismatch and unit-test.

Adds an error-resilient mode where frames can be continued
to be decoded even when there are errors (due to network losses)
on a prior frame. Specifically, backward updates are turned off
and probabilities of various symbols are reset to defaults at
the beginning of each frame. Further, the last frame's mvs are
not used for the mv reference list.

Change-Id: Ie6e2bcd436b1e8643270356d3a930e8989ff52a5