Adds an error-resilient mode with test.
Adds an error-resilient mode where frames can be continued
to be decoded even when there are errors (due to network losses)
on a prior frame. Specifically, backward updates are turned off
and probabilities of various symbols are reset to defaults at
the beginning of each frame. Further, the last frame's mvs are
not used for the mv reference list, and the sorting of the
initial list based on search on previous frames is turned off
as well.

Also adds a test where an arbitrary set of frames are skipped
from decoding to simulate errors. The test verifies that there
are no decoding errors for the remaining frames after the error
frames, as well as that the mismatch PSNR between the decoder's
version of the post-error frames and the encoder's version is at
least 20 dB.

Rebase.

Change-Id: Ie6e2bcd436b1e8643270356d3a930e8989ff52a5