Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

(6 inline comments)

I had the idea that rather than shutting off all the updates, we could save a copy of the entropy model with each reference frame. The idea is that if you're predicting from a frame you're calling "last" then it's ok to use its entropy model too. In combination with the reference frame updates I'm working on, this might be a more generic way of handling what we're already doing with copy of the FC we're already keeping for ARF. I think we'd still want to shut off the lastmv search, since keeping the motion field from all frames is way too big.

Agree with Jim that this needs a test.

Patch-set: 1
Label: Code-Review=-1
Label: Commit-Queue=0
Label: Verified=0