Update patch set 2
Patch Set 2:

Code wise OK in so far as it goes but simple removal of buffer copy hurts quality quite a lot in some cases. Most of this is down to loosing the copy of the old GF into the ARF buffer in bitstream.c. Even when coding using regular forward ARF updates this copy insured that there are two prior high quality buffers available during the coding of the next hidden arf update.

    if (pc->refresh_golden_frame)
      pc->copy_buffer_to_arf  = 2;

It should be possible to work around this and replicate the old behavior but this needs to be marked as a TODO so we don't forget...

Patch-set: 2
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0