Convert subpixel filters to use convolve framework
Update the code to call the new convolution functions to do subpixel
prediction rather than the existing functions. Remove the old C and
assembly code, since it is unused. This causes a 50% performance
reduction on the decoder, but that will be resolved when the asm for
the new functions is available.

There is no consensus for whether 6-tap or 2-tap predictors will be
supported in the final codec, so these filters are implemented in
terms of the 8-tap code, so that quality testing of these modes
can continue. Implementing the lower complexity algorithms is a
simple exercise, should it be necessary.

This code produces slightly different results in the EIGHTTAP_SMOOTH
case, since the filter is now applied even when the motion vector is 0.
The other filtering modes take advantage of the optimization where
filtering can be skipped in this case, and while it could be done for
this mode as well, there are disadvantages to doing so, so this can
be discussed at review time. Specifically, this gives the encoder
additional flexibility, and does not *require* that the shortcut be
implmented (eg, you can implmenent only a 2D convolver rather than
requiring 3 different cases to be implemented. It may be useful to
do so, to allow the filter to be applied in one direction only,
for example.

Full quality results will be forthcoming.

Change-Id: I7d469716091b1d89b4b08adde5863999319d69ff