Update patch set 1
Patch Set 1:

I did consider putting additional specializations in for the block size, but I think it's a premature optimization at this point. You can do the dispatch from the asm layer for the cost of a well-predicted branch, and it gives you a little more flexibility. For instance, you could specialize only 8px and 4px, and implement the 16px as a loop around the 8px, which you can't do if you have the width baked in to the function call (except by inserting a shim and calling a common function, but now you're back where you started with a branch in the asm layer function, or increased code size to duplicate the code).

It's not an obvious win, and not really hard to do later, so I think it's best to leave it out until we've optimized enough other things that it's going to make a measurable impact.

Patch-set: 1
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0