[WIP] Add column-based tiling.
This patch adds column-based tiling. The idea is to make each tile
independently decodable (after reading the common frame header) and
also independendly encodable (minus within-frame cost adjustments in
the RD loop) to speed-up hardware & software en/decoders if they used
multi-threading. Column-based tiling has the added advantage (over
other tiling methods) that it minimizes realtime use-case latency,
since all threads can start encoding data as soon as the first SB-row
worth of data is available to the encoder.

There is some test code that does random tile ordering in the decoder,
to confirm that each tile is indeed independently decodable from other
tiles in the same frame. At tile edges, all contexts assume default
values (i.e. 0, 0 motion vector, no coefficients, DC intra4x4 mode),
and motion vector search and ordering do not cross tiles in the same
frame.

Tile independence is not maintained between frames ATM, i.e. tile 0 of
frame 1 is free to use motion vectors that point into any tile of frame
0. We support 1 (i.e. no tiling), 2 or 4 column-tiles.

Resuls: with tiling disabled, we go up slightly because of improved edge
use in the intra4x4 prediction. With 2 tiles, we lose about ~1% on derf
and about 0.35% on HD. With 4 tiles, we lose another ~1.5% on derf and
another 0.77% on HD. Most of this loss is concentrated in the low-bitrate
end of clips, and most of it is because of the loss of edges at tile
boundaries and the resulting loss of intra predictors. STD/HD results
are pending.

TODO:
- more tiles (perhaps allow row-based tiling also, and max. 8 tiles)?
- maybe optionally (for EC purposes), motion vectors themselves
  should not cross tile edges, or we should emulate such borders as
  if they were off-frame, to limit error propagation to within one
  tile only. This doesn't have to be the default behaviour but could
  be an optional bitstream flag.
- ideally, the loopfilter becomes a 2-stage thing, where stage 1
  is independent for each tile and applies to the whole area within
  a tile only; then stage 2 (after all decoding/reconstruction/stage-
  1-loopfiltering is done) can optionally do deblocking at tile edges.
  This would allow decoding-interleaved, parallelizable loopfiltering.

Change-Id: I5951c3a0742a767b20bc9fb5af685d9892c2c96f