Update patch set 16
Patch Set 16:

Something else to consider: allow more tiles (8 columns) and enforce a maximum tile width (2K). If you want to encode more than 2K wide you have to use tiles. This provides reasonable bounds on the amount of memory required for hardware implementations, and was a request from the hardware guys when I was working on this.

Patch-set: 16