Butterfly ADST based hybrid transform
Devise a variant of the original ADST, which allows butterfly
computation structure. This new transform has kernel of the
form: sin((2k+1)*(2n+1) / (4N)). Its butterfly structure was
reported in Z. Wang, "Fast algorithms for the discrete W
transform and for the discrete Fourier transform", IEEE Trans.
on ASSP, 1984.

This patch includes the butterfly implementation of the inverse
ADST/DCT hybrid transform of dimension 8x8.

Change-Id: I3533cb715f749343a80b9087ce34b3e776d1581d