Update patch set 2
Patch Set 2:

re: 256 pixel min tile width. In hardware implementations, in-loop filter is performed pipelined with decoding (instead of a separate pass like you do in libvpx). Filter has to be applied also over tile boundaries. This means that the right edge of the previous tile will be read from DRAM when decoding and filtering current tile. 
-> Reading tile boundary causes memory bandwidth overhead
-> limiting minimum tile width to 256 (or greater) limits the amount of overhead. 

I would also argue that tiles are not needed for CIF or below resolutions, as those are trivial to decode with a single core CPU.

Patch-set: 2
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0