Update patch set 3
Patch Set 3:

@John: tiles align with 64x64 SBs, so alignment is not an issue. Power-of-two is purely practical, it makes the code somewhat simpler, and technically there isn't much difference between 8 vs. 9, whereas 2 vs 4 vs 8 vs 16 all give the same benefit (50%) of what you're looking at.

Why does this decrease latency and a bool coder doesn't? Because we need the tile size as part of the bitstream, so we can only send a tile once the tile is finished. Here, that's at 25% of the frame, it used to be 100%, so 75% reduction in latency from this end. It's true that slices can help there too, but they have their own issues. This is certainly simpler from an implementation point of view. Also, slices don't allow tile columns unless you have multiple communication channels for each tile column.

Patch-set: 3
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0