Update patch set 3
Patch Set 3:

Sure. For low latency streaming in WebM, we need to arrange the frame such that it never requires backtracking to write it out. Each byte needs to be written once, and you can never move the write pointer backwards. This is a very useful property, as it lets you stream over FIFOs, and you can plug in adapters that take the compressed bitstream and transform it easily (ie, call a function that adds FEC as you write bytes out). This is wholly unrelated to tiling though (other than it needs to obey these properties). You don't need tiling to get low latency if you arrange the stream in this way, so I see fixing this as a separate patch.

This patch is about 2d tiling, which is very useful, if the tiles are independent. The only reason to add offsets into the bitstream is to make parts of it random access, and it's not all that useful to get random access to a dependent tile. It'd let you pack the dependent parts non-contiguously, or work in raster rather than column order but I'm not clear on if that's what you're trying to achieve or why that's important. On the other hand, independent tiles bring a whole bunch of functionality, and seem easy to implement on top of this patch. That's an an orthogonal issue to whether they're necessary or useful wrt latency.

Patch-set: 3