Update patch set 3
Patch Set 3:

The use case with partially decoding a high resolution still image is valid and justifies independent row tiles. I'm ok with that. It's true that the latency issue can be solved by getting rid of carry propagation as John proposed, that would be a good thing from hw perspective.

Patch-set: 3