Update patch set 5
Patch Set 5:

This is sort of a precursor to and overhaul of the coefficient context with changes to the number of coef bands and the number and perhaps meaning of the pt bands.

As things stand here the use of different coefficient banding arrays here does not give much benefit and does give rise to large fixed structures and the need to pass pointers to them around. Changing the number of each type of band then requires structures like the vp9_coeff_probs default_*[] to be re-dimensioned and maintaining multiple copies of these huge and unwieldy structures under an experiment is messy.... 

If it really bothers you I can make this patch so that it exactly matches the behaviour we had before but with smaller data structures and the selector based on tx size, but reversion to the old behaviour is not difficult anyway and in particular, while we are looking at things like removing Y2 and perhaps using intra/inter as a table selection criteria rather than hybrid/non-hybrid I would rather make sure we don't have multiple copies of the default tables.

Patch-set: 5
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0