Update patch set 3
Patch Set 3: Looks good to me, approved

I think that this is an interesting discussion but goes beyond what HAS to be done in THIS patch.

We can come back to this later but I think it is OK to submit this as is and get it merged.

Patch-set: 3
Reviewer: Gerrit User 1000899 <1000899@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+2
Label: Commit-Queue=0
Label: Verified=0