Update patch set 2
Patch Set 2:

So, this changes each and every filter function to enable scaling, that seems like:
A) in the encoder case where we have a scaled reference frame, this will add a _HUGE_ overhead (since for every frame where that frame is available, it will test all modes - including a full motion search - with scaling). Shouldn't the reference be scaled and then be used in the RD as-such?
B) In the (somewhat unlikely) case that you do want to scale, does that really need to be exposed as such? The RD loop shouldn't have to care that it scales or not, it should call a filter function that just does the right thing. Exposing specific implementation details (8tap, smooth, sharp, scale) shouldn't be done at that level.

Patch-set: 2
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0