Update patch set 2
Patch Set 2:

To refine what I said before about the motion search, there are cases where it'd be better to scale inside the motion search loop. If the reference frame is smaller, then you'd want to scale the source block down (once), and search inside the unscaled smaller reference frame, then scale the resulting MV back up. This would be faster than a normal, unscaled motion search.

When the reference is larger than the current frame, the search is infrequent (ie, that it's your last resort predictor), while the candidate predictors are non-overlapping, it should be faster to scale inside the loop. When the search radius is small it'd be better to scale one block that contains all MVs and search within that. But if you were going to search on every block (say its the last frame), it'd be better to scale the whole frame to a temporary area to use when searching, which you can then of course save the scaled result to use later.

In the case of spatial scalability where the reference is temporally aligned with the current frame, then you probably can skip the motion search entirely and use 0,0. In the case where you're switching resolutions, then you're probably scaling only once every few frames, and you could save the scaled result so you only do it once. So in any case I don't see anything about this patch that dooms us to poor performance always.

What I was driving at before is that I probably won't try to do any of these tricks at first, just get something working. I don't yet know whether I'll do it  up front by scaling the whole frame, or do it inside the loop. I'll probably take the shortest path to correct code. There's no one-size-fits-all solution on the encoder side I don't think, and there are all sorts of optimizations that can be done. I don't think they affect the decoder, which is where I'm focusing primarily at the moment.

Patch-set: 2