Update patch set 2
Patch Set 2: Looks good to me, but someone else must approve

I think this is OK, but Debargha, can you double check that the removal of the mv entropy stats is indeed warranted? (I.e. either we don't need it or it's hard to fix.)

Patch-set: 2
Label: Code-Review=+1
Label: Commit-Queue=0