Update patch set 2
Patch Set 2:

So Paul, for the default tables you generated, you can still reuse them; remove the ones at plane type index [0] and [1], and switch [2] and [3] around (so Y-with-DC comes first, followed by UV). Remove the old [0]/[1] (Y-no-DC and Y2), and you can omit the ones at 32x32 [1] and all hybrid [1], since the UV are not used for them.

I'll try to re-run them periodically over the next few weeks as we experiment more with the coefficients.

Patch-set: 2