Update patch set 1
Patch Set 1: Looks good to me, approved

Agree that this is a release blocker, and your approach LGTM. I don't think this needs to be done as an experiment, we should just make it the default. I'll LGTMA this and you can either merge as is and remove the old code in a separate commit, or you can post a new patch set on this change that does the same and we can get a wider review.

If you can't submit and want to take the former approach, ping me and I'll submit this for you.

Patch-set: 1
Label: Code-Review=+2
Label: Commit-Queue=0
Label: Verified=0