Update patch set 2
Patch Set 2:

Can you be more specific wrt the additional requirements imposed upon other implementations? Being "not super-happy" doesn't really give me anything to respond to.

I acknowledge that this is more complicated to implement than the simplest thing possible, but not overly so. In general, I'm trading implementation complexity for significantly lower computational complexity, and that seems appropriate. There may some additional changes like going to 1/32 subpel positions that could let us simplify some of the offset updates and might improve quality of the scaled predictor, but the core idea seems sound, so it's not obvious why this is so objectionable.

Patch-set: 2