Update patch set 2
Patch Set 2:

I didn't specifically say objectionable, I just said I'm a little unhappy with the implementation complexity. I do honestly think this can be done simpler. If the quality or speed are indeed greatly enhanced by this implementation, that may be worth it, but we currently have no evidence for that yet.

In fact, I still believe the encoder will be slower, because it scales twice. Also, since the dummy frame for motion search isn't the actual target, there will be a quality impact (inaccurate motion search). I've also voiced a concern on the design of the downscaling filter coefficients (the period/wavelength thing), generally you increase filtersize during downscaling because the source pixel coverage per target pixel increases as the downscaling level increases; this may also affect quality.

So that leaves that maybe the decoder is faster.

This is a massively complex set of patches, which touches every single part of the encoder and adds a lot of complexity to each part of the encoder, which is a big cost given the benefits over a much simpler implementation. That's what I'm a little unhappy about. I'm not saying I find it objectionable, just that I'm a little unhappy about it.

I don't think there's much that can be done about that, I've tried giving more addressable feedback, which we should probably focus on so this can be merged.

Patch-set: 2