Update patch set 2
Patch Set 2:

WRT implementation complexity, I'm talking about algorithmic complexity gains. It's like saying quicksort is too complex to implement, so we should use bubblesort. There are cases where bubblesort is faster than quicksort, but you need to have a good reason to believe your workload exploits that path before preferring the theoretically worse version. That said, I love simplifying, and if you can suggst a simpler implementation with the same algorithmic complexity, I'd be happy to look at that. I believe that my approach is unconditionally as good or better than the simple approach of scaling unconditionally:

  Scaling, Subpel:       max(F, S) <  F + S
  No Scaling, Subpel:    F         <  F + S
  Scaling, No Subpel:    S         == S
  No Scaling, No Subpel: 0*        <  S

  * not really 0 (copy), but equal to a constant that's also present in all other cases

I haven't heard any evidence that this won't hold true in the real application, or that the algorithmic complexity doesn't matter in this case, or that I've made some fatally flawed assumption. I wrote a test app that does a full prediction/scale both ways as would happen on the decoder, and the gains are obvious (times in usec, 50 frames 720p<->1080p, F == S):

  upscaling 256/256 mbs
  scale_and_subpel: 2643881
  scale_then_subpel: 3307552
  downscaling 256/256 mbs
  scale_and_subpel: 1671811
  scale_then_subpel: 1747064

  upscaling 192/256 mbs
  scale_and_subpel: 2062356
  scale_then_subpel: 2742751
  downscaling 192/256 mbs
  scale_and_subpel: 1118820
  scale_then_subpel: 1846113

  upscaling 128/256 mbs
  scale_and_subpel: 1273704
  scale_then_subpel: 2763321
  downscaling 128/256 mbs
  scale_and_subpel: 642289
  scale_then_subpel: 1748330

  upscaling 64/256 mbs
  scale_and_subpel: 574175
  scale_then_subpel: 2856640
  downscaling 64/256 mbs
  scale_and_subpel: 649559
  scale_then_subpel: 1974391

  upscaling 1/256 mbs
  scale_and_subpel: 160831
  scale_then_subpel: 2831036
  downscaling 1/256 mbs
  scale_and_subpel: 67927
  scale_then_subpel: 1797847

I'll make the code available so you can play with it. The cheaper scaling is the less this optimization matters, but I think we can consider it sufficiently large that whatever overhead the infrastructure supporting this mode adds will be dwarfed by the time we spend in these functions. If it's say F, or even F/4, then this is probably 10s of percent savings in the overall decode time, given how much time we spend in F in the decoder. Enough that it might be the difference between working and not on an embedded device.

WRT the encoder speed, as I said before, I did the fastest thing to implement that works. There's unlimited time to make it better later. The degenerate case is that you have to downscale and don't have any other references to use where you'd effectively scale twice so you can do motion search. Even in that case, you only have to do that on one frame, or maybe at most 3-4 over a second if you're doing temporal layers. There are all sorts of tricks you can play to make this cheaper in the more typical cases.

WRT the dummy frame, right now I think it is equivalent to the actual target for non-SPLITMV modes. It might be equivalent for SPLITMV too, but I can't convince myself at the moment. In any case, I think it's very close.

WRT downscaling and filter design, my point was that you can do good 2:1 downscaling with 8 taps, since it includes pixel [2] and two on either side of it. You'd typically double the number of taps to go to 2:1, so it'd take 16 taps to map our 8 tap filter exactly, but you could do a 4 tap bicubic kernel expanded to 2:1 in 8 taps, for example. Equivalently, you can move the input data pointer and use the 4 taps directly, which is the approach I'm taking right now. What I'm doing only gives an approximation, but I haven't noticed any visual artifacts, so it's either probably close enough or I haven't found the ratio that breaks it. We can definitely do something more mathematically accurate for both upscaling and downscaling that generates the correct filters to use based on the scaling ratio on a per-frame basis. Whether we /must/ do so is unclear, but if we want to, it plugs into this framework trivially and can be added later. Needs more testing when I have a better test harness.

WRT this only being good for the decoder, I think this approach is helpful on the encoder too. There might be some degenerate cases, but in general, the idea is to scale only when you need it, and only the data that needs to be scaled, and that should be faster than scaling the whole frame unconditionally and only reading a perhaps small subset of the scaled frame. In addition, I expect this will be used more in realtime modes than in best modes where all references are searched more exhaustively.

WRT this touching a lot of code in the encoder, I think that's mostly a factor of the encoder's design. At its core, this is a pretty simple change, and shouldn't get in the way of additional cleanups/refactoring that would narrow its impact. I don't see it as particularly complex or hard to maintain. Net code growth in vp9/encoder for this patch series is only about 200 lines, probably similar to what a simpler approach would be, though there'd have been less churn.

I'll incorporate your comments as I've previously responed to them. Please -1 those updated patches if you're not happy with the resolution rather than leaving unscored, so I know where this patch series stands.

Patch-set: 2