Update patch set 3
Patch Set 3: (1 inline comment)

I confess to some residual anxiety in regard to the very generalized approach to the problem. In particular the level of complexity and the widespread and invasive nature of the changes required.

The fact that in many situations we need now to account for and deal with the possibility that the compression scale and that of the reference buffers may be different, plus reference buffers may differ from each other (e.g. in the various compound modes inter/inter and inter/intra) seems to me to open a large number of new opportunities for bugs and problems.

What you have done so far seems solid but the changes are, as I feared quite complex and widespread.

Patch-set: 3
Reviewer: Gerrit User 1000899 <1000899@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0
Label: Verified=0