Update patch set 3
Patch Set 3: (3 inline comments)

I think a configure flag is appropriate for code that we might want to remove later, or we need to isolate from other people working in the code. With this commit in particular, I think it's more about refactoring that we wouldn't want to remove anyway (eg combining the 1st/2nd code everywhere) than it is about scaling. I also think we want to get this into as many code paths as possible as quickly as possible to both to ferret out bugs and to make sure people think about how their features interact with this one, so I'd rather not make it opt-in. Experiments are a useful tool in many cases though, and if there are specific chunks in future patches that look like they really ought to be protected with an experiment, then definitely make that comment in those places. But I'd like the majority of this commit series to be committed without being behind a flag.

Patch-set: 3
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0