Make superblocks independent of macroblock code and data.
This commit basically makes sb32/64 blocks independent of mb-related
blocks (except for the RD, where it doesn't matter). This allows us
to completely separate the Y, U and V tokens, implement the 16x16
U/V transform for 16x16 Y transform blocks (previously it was only
enabled if the Y used 32x32) in 32x32 superblocks, and the 32x32
U/V transform for 32x32 Y transform blocks in 64x64 superblocks
(previously it used the 16x16), allows trellis optimization of 32x32
Y and U/V transform blocks, and makes superblocks feel less like it
was a sticky hack on top of macroblocks.

The default probabilities for 32x32 U/V are currently a copy of the
16x16; real probabilities will be generated over the next few days.

Change-Id: Ibad7c3ddfe1acfc52771dfc27c03e9783e054430