Add optional filtering to intra predictors
This patch adds the option to filter each of the intra-frame
predictors for both intra- and inter-frames.

The filter is not applied to DC_PRED & TM_PRED modes and is
currently not applied to 4x4 blocks.

The prediction filter is signaled off/on at the MB level for
the Y & UV components. Probabilities for each flag are
encoded at the frame level.

Later patches will improve the entropy coding, add a frame
level flag to enable/disable filtering completely, and add
filtering to 4x4 blocks (using a shorter filter).

Change-Id: I71ccaa0c48072b08617ec5cfd50f409fe1fe4278