Update patch set 2
Patch Set 2:

I think we need to fix the problem at it's source, which is in update_rd_ref_frame_probs, as you mention.

Clipping the prob values at the 2 places where relative adjustments are made (+40 and -20) seems to me like the right thing to do.

Whilst the traps that you add are a sensible safeguard, I would argue that since the function is only called from two areas of the code, the above fix should ensure that there is no danger of outranging now.

Patch-set: 2
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0