Implicit weighted prediction experiment
Adds an experiment to use a weighted prediction of two INTER predictors,
where the weight is one of (1/4, 34), (1/2, 1/2), or (3/4, 1/4), and
is chosen implicitly based on consistency of the predictors to the already
reconstructed pixels to the top and left of the current macroblock
or superblock.

Currently the weighting is not applied to color or SPLITMV modes,
which default to the usual (1/2, 1/2) weighting.

Change-Id: I5438539ff4485c5752874cd1eb078ff14bf5235a