Implicit weighted prediction experiment
Adds an experiment to use a weighted prediction of two INTER predictors,
where the weight is one of (1/4, 3/4), (3/8, 5/8), (1/2, 1/2), (5/8,
3/8) or (3/4, 1/4), and is chosen implicitly based on consistency of the
predictors to the already reconstructed pixels to the top and left of the
current macroblock or superblock.

Currently the weighting is not applied to SPLITMV modes, which default to
the usual (1/2, 1/2) weighting. The same weighting is used for Y and UV
components, where the weight is derived from analyzing the Y component
only.

Results (over compound inter-intra experiment)
derf: +0.18%
hd: +0.49%
yt: +0.34%

The experiment suggests bigger benefit for explicitly signaled weights.

Change-Id: I5438539ff4485c5752874cd1eb078ff14bf5235a