Modeling default coef probs with distribution
Replaces the default tables for single coefficient magnitudes with
those obtained from an appropriate distribution. The EOB node
is left unchanged. The model is represented as a 256-size
codebook where the index corresponds to the probability of the
zero node on the coef encoding tree. The node probabilities
are then obtained by looking up the codebook using the prob of
the zero node as the index. The main advantage is that the
default prob tables will become considerably smaller and manageable,
Besides there is considerably less risk of over-fitting for a
training set.

Currently forward updates  are optionally (controlled by a macro)
restricted to only convey probabilities from the codebook. However
backward updates are left unconstrained.

Various distributions are tried and the one that gives the best
results is the Generalized Gaussian with shape parameter 0.75.
The results are quite close to fully trained tables with the max
difference being -0.265% for hd, and yt even improving by +0.206%.

Results for unconstrained forward update (Generalized Gaussian 0.75):
derf: -0.080%
yt: +0.206%
stdhd: -0.133%
hd: -0.265%

Results for constrained forward update (Generalized Gaussian 0.75):
pending.

Change-Id: Iae050f3cbcc6d8b3f204e8dc395ae47b3b2192c9